Refactor NFT endpoints to nfts.endpoints.ts
